+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - Add: `wrangle.normalise_volume` to deal with different volume formats of Uniswap v2 and v3 
 - Add: Support for Coingecko data loading and metadata cross referencing. See `tradingstrategy.alternative_data.coingecko`.
 - Add: `tradingstrategy.alternative_data.coingecko.categorise_pairs()` to tag trading pair data with their CoinGecko category
-- 
+- Add: `deduplicate_pairs_by_volume()` to make it easier to construct trading pair baskets from open-ended universes 
+
 # 0.24.3
 
 - Fix: Allow to run without Jupyter notebook/IPython installed. Make sure you use `Client.create_live_client()` instead of `Client.create_jupyter_client()`